requirements.txt
numpy==1.26.1
scikit-learn==1.5.2
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.11.2
apache-airflow
pandas
tqdm
psycopg2-2.9.10
sqlalchemy